Convert to use int64_t to aovid int oout-of-range
BUG=aomedia:518

+++ av1/encoder/mcomp.c
@@ -2346,9 +2346,9 @@ int av1_full_pixel_search(const AV1_COMP *cpi, MACROBLOCK *x, BLOCK_SIZE bsize,
 /* checks if (r, c) has better score than previous best */
 
 #define MVC(r, c)                                                         \
-  (mvcost                                                                 \
+  (unsigned int)(mvcost                                                   \
        ? ((mvjcost[((r) != rr) * 2 + ((c) != rc)] + mvcost[0][((r)-rr)] + \
-           mvcost[1][((c)-rc)]) *                                         \
+           (int64_t)mvcost[1][((c)-rc)]) *                                \
               error_per_bit +                                             \
           4096) >>                                                        \
              13                                                           \
@@ -2936,9 +2936,9 @@ int av1_masked_full_pixel_diamond(const AV1_COMP *cpi, MACROBLOCK *x,
 
 /* checks if (r, c) has better score than previous best */
 #define MVC(r, c)                                                         \
-  (mvcost                                                                 \
+  (unsigned int)(mvcost                                                   \
        ? ((mvjcost[((r) != rr) * 2 + ((c) != rc)] + mvcost[0][((r)-rr)] + \
-           mvcost[1][((c)-rc)]) *                                         \
+           (int64_t)mvcost[1][((c)-rc)]) *                                \
               error_per_bit +                                             \
           4096) >>                                                        \
              13                                                           \
